add Statistic Service+Controller

+++ individualTask/src/main/java/org/teamscore/individualTask/models/DTO/statistic/CategoryStatistic.java
@@ -3,17 +3,17 @@ package org.teamscore.individualTask.models.DTO.statistic;
 import java.math.BigDecimal;
 
 public class CategoryStatistic extends AbstractStatistic {
-    private final float percentage;
+    private final double percentage;
     private final String color;
 
 
-    public CategoryStatistic(String name, BigDecimal totalSum, float percentage, String color) {
+    public CategoryStatistic(String name, BigDecimal totalSum, double percentage, String color) {
         super(name, totalSum);
         this.percentage = percentage;
         this.color = color;
     }
 
-    public float getPercentage() {
+    public double getPercentage() {
         return percentage;
     }
